Examples moved up
