+++ BlockIo.UnitTests/KeyTest.cs
@@ -2,7 +2,7 @@
 using NUnit.Framework;
 using System;
 
-namespace BlockIo.UnitTests
+namespace BlockIoLib.UnitTests
 {
     public class KeyTest
     {
