+++ BlockIo/Program.cs
@@ -1,18 +1,10 @@
-﻿using BlockIo.Examples;
-using System;
-
-namespace BlockIo
+﻿
+namespace BlockIoLib
 {
     class Program
     {
         static void Main(string[] args)
         {
-            /**
-             * Examples
-             */
-
-            Basic basic = new Basic();
-            basic.RunBasicExample();
         }
     }
 }
