+++ BlockIo/ResponseObject/BlockIoResponse.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace BlockIo
+namespace BlockIoLib
 {
     public class BlockIoResponse <T>
     {
