--- BlockIoLib.UnitTests/data/sign_and_finalize_dtrust_withdrawal_request.json
@@ -1,3 +1,0 @@
-﻿{
-  "signature_data": "{\"reference_id\":\"29c29da0e2fe09dc911cb85162f75dcc91d8edf0ad902f1a1327238d6a97ba85\",\"inputs\":[{\"input_no\":0,\"signatures_needed\":3,\"data_to_sign\":\"d3dea799a6104424582e570fba81417c8f204b898306d9e877884a3ecc5dd977\",\"signers\":[{\"signer_address\":\"mh6DXT3EXNvcjsphQtaFNDf71fBazMkyaJ\",\"signer_public_key\":\"02510e29d51e9a4268e6a5253c1fbd8144857945b82acb0accfc235cc7ca36da11\",\"signed_data\":\"304402205e145909ffd852200a118b151393d7b49b1024c33d8d54e469eb7fc9359130d60220711577ab007feaf90f823d045ceb72db9788e63604e336fdd43f0f7d4b0e665a\"},{\"signer_address\":\"mpS4XHVzeXEXnMcwhZHTBUHJTT8ELTJgb6\",\"signer_public_key\":\"0201a819bf549c253c4397bdd1535374de39e9bc278f637afdc27642d52cf79139\",\"signed_data\":\"3045022100d3fb7c071b158204ad2cebbbfe4193e5acdbb7e624691466505a0bbbef69e2f0022001f7ef283d92aeda19dd627d428b4eb32597440e52a75c85d7e805c3554766da\"},{\"signer_address\":\"mgiiQLxi26mYLagyECR9jJF9UxWk3DqSfu\",\"signer_public_key\":\"039e3aa9ea182ccdaff2d8d150010b27cc4765c1d55ce674e52631af7376354d62\",\"signed_data\":\"304402207d106e86dc9558ddb3413aa297ebc49283404566229566a8f82460d5d5c889750220129c676fd23fac51cfc7a2fed0bae071fc8d2e8069067c77d3b3ab8616fd0a7d\"},{\"signer_address\":\"n1SoEoc2Fn4VamANqj4sd3taadHr4bmMNn\",\"signer_public_key\":\"03ee980e6334142342fcd9e6facfecfa139981e2276584c91d6a9739d533ac99fc\",\"signed_data\":null}]},{\"input_no\":1,\"signatures_needed\":3,\"data_to_sign\":\"ee0ca83531b1a117d63f70dbe6ce147305fe00e70810e4c9a8757208473c69b5\",\"signers\":[{\"signer_address\":\"mh6DXT3EXNvcjsphQtaFNDf71fBazMkyaJ\",\"signer_public_key\":\"02510e29d51e9a4268e6a5253c1fbd8144857945b82acb0accfc235cc7ca36da11\",\"signed_data\":\"3045022100cb1cb6e0a4fa7321abbfe256bd502e78d712a9ec107f0a84e8198cfbe14a518902201dab625d521be5aa114816ca05fd5fa4fcbe5205be0352a37f8439e1f712b8c6\"},{\"signer_address\":\"mpS4XHVzeXEXnMcwhZHTBUHJTT8ELTJgb6\",\"signer_public_key\":\"0201a819bf549c253c4397bdd1535374de39e9bc278f637afdc27642d52cf79139\",\"signed_data\":\"3044022024900bc8beb34afdc9b4186bd3d6a10cc534e72378684385061d23f90a9eaa33022062ce52b7e7632cedb6b67d9884d7f2107691d2f2290feb6d277ad53736136a5b\"},{\"signer_address\":\"mgiiQLxi26mYLagyECR9jJF9UxWk3DqSfu\",\"signer_public_key\":\"039e3aa9ea182ccdaff2d8d150010b27cc4765c1d55ce674e52631af7376354d62\",\"signed_data\":\"3045022100f76a2351894413075b1e62ea08f74b044d61f9d7368921aee583dd6f8c900d350220462549b05f7db071bbd0899ea513057c39729424d75238d969bda0b5c352b1e3\"},{\"signer_address\":\"n1SoEoc2Fn4VamANqj4sd3taadHr4bmMNn\",\"signer_public_key\":\"03ee980e6334142342fcd9e6facfecfa139981e2276584c91d6a9739d533ac99fc\",\"signed_data\":null}]}]}"
-}
