--- BlockIoLib.UnitTests/data/sign_and_finalize_sweep_request.json
@@ -1,3 +1,0 @@
-﻿{
-  "signature_data": "{\"reference_id\":\"1139ee242378134abbe6e32fc5afb8b2a445a9ac69fed80e4dd6036312b05461\",\"inputs\":[{\"input_no\":0,\"signatures_needed\":1,\"data_to_sign\":\"4187096c9d66f5ff9f6423d76168cf66136a62d982ecdbebe08f153873343757\",\"signers\":[{\"signer_public_key\":\"02d9798bba9d2c81ba38f4f3b6f586c199a21b2774abafb07f095478ed13bfda61\",\"signed_data\":\"30440220458328ffe90c42aa0c1baa40cc1c7eebd1347876398839fc1a91aa4d69e7d088022011fce980dd840202a3e38bdb8a1863932bc4a6aa362b920cbb090d4ad56071b5\"}]},{\"input_no\":1,\"signatures_needed\":1,\"data_to_sign\":\"6877f9a559d9237050fd369114ec5a8f897f2c9f2aabf3c7d8d87dbf0aff197b\",\"signers\":[{\"signer_public_key\":\"02d9798bba9d2c81ba38f4f3b6f586c199a21b2774abafb07f095478ed13bfda61\",\"signed_data\":\"3044022048b9de46a460bc0add1694b2b4596015e45097e493f5076153f0ff252dc40eea022012615778a8ec66598476ffbeb1bea4fb55b7f40a8d812f2c87b2969621e20a0d\"}]}]}"
-}
