--- BlockIoLib.UnitTests/data/sign_and_finalize_withdrawal_request.json
@@ -1,3 +1,0 @@
-﻿{
-  "signature_data": "{\"reference_id\":\"25cd8b0ad5cf8987c99e16921149cdb0680b1c1360276b4d652904958b19a8bd\",\"inputs\":[{\"input_no\":0,\"signatures_needed\":1,\"data_to_sign\":\"98d864802376e9eaf5115d970303d4dc5c1d2557dc93568ea72e7b31692867e5\",\"signers\":[{\"signer_address\":\"ng8SXuwBoPk7A8MXsNBMS6DMkFNNPCSZdL\",\"signer_public_key\":\"029774ca6d07e17faeac23f4075b680ddc8ee24d3214980e79ca5a5842f6828ff1\",\"signed_data\":\"3044022065e67caa5b14cff32aa9a420ebe63dcba7610e41cabba3ba4619996149a555fc022076c043cceaa5fbde7ee09603416798c569939c360d839991b93e5790657b48b3\"}]},{\"input_no\":1,\"signatures_needed\":1,\"data_to_sign\":\"453b24a3ff7728392c9be8178ef5d66cd0e6be8a2726fc182eaab0546e6a3206\",\"signers\":[{\"signer_address\":\"ng8SXuwBoPk7A8MXsNBMS6DMkFNNPCSZdL\",\"signer_public_key\":\"029774ca6d07e17faeac23f4075b680ddc8ee24d3214980e79ca5a5842f6828ff1\",\"signed_data\":\"304402203543cdd59db831237de82e19551d184b1acc2de868574c0ba920b7eaf9f26afd0220273df0976046ab5a0bd08e47fae599003d2145bfb2b04162335d65db22fb0744\"}]},{\"input_no\":2,\"signatures_needed\":1,\"data_to_sign\":\"7bdcbb1a0c0d257fecfe638499f06fae6c686f7103595703330782cd6bf64f52\",\"signers\":[{\"signer_address\":\"ng8SXuwBoPk7A8MXsNBMS6DMkFNNPCSZdL\",\"signer_public_key\":\"029774ca6d07e17faeac23f4075b680ddc8ee24d3214980e79ca5a5842f6828ff1\",\"signed_data\":\"3044022057da6e08079082cea2ebd182c1247c1e7c0f03275437788b62d3331b9b615bfc02200dcc811419e8d3be5d06487e253c53fb5fd441b2b37745c22163e322ea94f82a\"}]}]}"
-}
