--- BlockIoLib/BlockIo.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Library</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Base58Check" Version="0.2.0" />
-    <PackageReference Include="dotenv.net" Version="2.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.Cryptography.KeyDerivation" Version="3.1.6" />
-    <PackageReference Include="NBitcoin" Version="5.0.49" />
-    <PackageReference Include="RestSharp" Version="106.11.4" />
-    <PackageReference Include="WireMock.Net" Version="1.2.17" />
-  </ItemGroup>
-
-</Project>
