--- BlockIoLib/Program.cs
@@ -1,18 +1,0 @@
-﻿using BlockIoLib.Examples;
-using System;
-
-namespace BlockIoLib
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            /**
-             * Examples
-             */
-
-            Basic basic = new Basic();
-            basic.RunBasicExample();
-        }
-    }
-}
