--- BlockIoLib/ResponseObject/BlockIoResponse.cs
@@ -1,9 +1,0 @@
-﻿
-namespace BlockIoLib
-{
-    public class BlockIoResponse <T>
-    {
-        public string Status { get; set; }
-        public T Data { get; set; }
-    }
-}
