+++ Examples/MaxWithdrawal/Program.cs
@@ -1,42 +1,37 @@
-﻿using dotenv.net;
+﻿using BlockIoLib;
+using dotenv.net;
 using dotenv.net.Utilities;
 using System;
 using System.IO;
 using System.Text;
 
-namespace BlockIoLib.Examples
+namespace MaxWithdrawal
 {
-    class MaxWithdrawal
+    class Program
     {
-        private BlockIo blockIo;
-        private EnvReader envReader;
-
-        public MaxWithdrawal()
+        static void Main(string[] args)
         {
             var path = Path.Combine(Directory.GetCurrentDirectory());
             path = Path.GetFullPath(path) + "/.env";
             DotEnv.Config(true, path);
             DotEnv.Config(true, path, Encoding.Unicode, false);
-            envReader = new EnvReader();
+            EnvReader envReader = new EnvReader();
 
-            blockIo = new BlockIo(envReader.GetStringValue("API_KEY"), envReader.GetStringValue("PIN"));
-        }
+            BlockIo blockIo = new BlockIo(envReader.GetStringValue("API_KEY"), envReader.GetStringValue("PIN"));
 
-        public void RunMaxWithdrawalExample()
-        {
             var balance = blockIo.GetBalance().Data.available_balance;
-            
+
             Console.WriteLine("Balance: " + balance);
 
             while (true)
             {
-                var res = blockIo.Withdraw(new { to_address= envReader.GetStringValue("TO_ADDRESS"), amount=balance.ToString() });
+                var res = blockIo.Withdraw(new { to_address = envReader.GetStringValue("TO_ADDRESS"), amount = balance.ToString() });
                 double maxWithdraw = res.Data.max_withdrawal_available;
 
                 Console.WriteLine("Max Withdraw Available: " + maxWithdraw.ToString());
 
                 if (maxWithdraw == 0) break;
-                blockIo.Withdraw(new { to_address= envReader.GetStringValue("TO_ADDRESS") , amount=maxWithdraw.ToString() });
+                blockIo.Withdraw(new { to_address = envReader.GetStringValue("TO_ADDRESS"), amount = maxWithdraw.ToString() });
             }
 
             balance = blockIo.GetBalance().Data.available_balance;
